big refactor, creating new folders and methods, move functionality to new methods
